Extracts case detail link
